Move post-condition verification from `stackslib` -> `clarity`
